requirements.txt
apify-client>=1.8.0
openai>=1.50.0
pyyaml>=6.0
python-dotenv>=1.0
structlog>=24.0
mediapipe>=0.10.0
opencv-python>=4.8.0
